[FE] team-03브랜치에서 dev-FE로 최신화

+++ BE/src/main/java/sidedish/com/domain/DiscountPolicy.java
@@ -8,9 +8,9 @@ import lombok.Getter;
 public class DiscountPolicy {
 
 	private String policyName;
-	private float discountRate;
+	private long discountRate;
 
 	public long calculateFixedPrice(long originalPrice) {
-		return (long) (originalPrice * ((100 - discountRate) / 100));
+		return (long) (originalPrice * ((100 - discountRate) / (double) 100));
 	}
 }
